Kiosk_website/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
pytz==2023.3
